ml_service/requirements.txt
django
djangorestframework
django-cors-headers
numpy
pandas
scikit-learn
pillow
requests
